refactor: DTO의 구조 정리로 인한 import문 변경
