+++ src/main/java/roomescape/controller/ReservationController.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.Reservation;
-import roomescape.dto.request.ReservationCreateRequest;
-import roomescape.dto.response.ReservationResponse;
+import roomescape.dto.reservation.request.ReservationCreateRequest;
+import roomescape.dto.reservation.response.ReservationResponse;
 import roomescape.service.ReservationService;
 
 @RestController
