+++ src/main/java/roomescape/dto/time/response/TimeCreateResponse.java
@@ -1,0 +1,4 @@
+package roomescape.dto.time.response;
+
+public class TimeCreateResponse {
+}
