+++ src/main/java/roomescape/service/ReservationService.java
@@ -6,8 +6,8 @@ import java.time.LocalTime;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.domain.Reservation;
-import roomescape.dto.request.ReservationCreateRequest;
-import roomescape.dto.response.ReservationResponse;
+import roomescape.dto.reservation.request.ReservationCreateRequest;
+import roomescape.dto.reservation.response.ReservationResponse;
 import roomescape.error.ErrorMessage;
 import roomescape.error.exception.InvalidValueException;
 import roomescape.repository.ReservationDAO;
